modificaciones

+++ src/components/Usuarios.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import { getFirestore, collection, getDocs, doc, updateDoc, deleteDoc } from "firebase/firestore";
+import { collection, getDocs, doc, deleteDoc } from "firebase/firestore";
 import Nav from '../partials/Nav';
-import { auth, firestore } from '../firebase'
+import { firestore } from '../firebase'
 import Footer from '../partials/footer'
 
 function Usuarios() {
@@ -18,14 +18,18 @@ function Usuarios() {
     };
 
     const handleEliminarUsuario = async (uid) => {
-        try {
-            const usuarioRef = doc(usuariosCollection, uid);
-            await deleteDoc(usuarioRef);
+        const confirmed = window.confirm("¿Estás seguro de que quieres eliminar este usuario?");
 
-            const nuevosUsuarios = usuarios.filter((usuario) => usuario.uid !== uid);
-            setUsuarios(nuevosUsuarios);
-        } catch (error) {
-            console.error("Error al eliminar el usuario:", error);
+        if (confirmed) {
+            try {
+                const usuarioRef = doc(usuariosCollection, uid);
+                await deleteDoc(usuarioRef);
+    
+                const nuevosUsuarios = usuarios.filter((usuario) => usuario.uid !== uid);
+                setUsuarios(nuevosUsuarios);
+            } catch (error) {
+                console.error("Error al eliminar el usuario:", error);
+            }
         }
     };
 
@@ -52,43 +56,62 @@ function Usuarios() {
             });
     }, []);
 
+    
+
+
+
     return (
 
         <div className="page-container">
             <Nav />
             <div className="usuarios-container">
                 <div className="container">
-                    <button className="botonnuevo">
-                        <Link className="btn btn-outline-secondary" to="/register">
-                            Agregar Nuevo Usuario
-                        </Link>
+                    <div className="container">
+                        <div className="d-flex justify-content-end mb-3">
+                            <Link className="btn btn-outline-secondary" to="/register">
+                                Agregar Nuevo Usuario
+                            </Link>
+                        </div>
+                    </div>
 
-                    </button>
-                    <table className="table table-hover">
-                        <thead>
-                            <tr>
-                                <th scope="col" className='text-uppercase'>Nombre</th>
-                                <th scope="col" className='text-uppercase'>Correo</th>
-                                <th scope="col" className='text-uppercase'>Cargo</th>
-                                <th scope="col" className='text-uppercase'>Rol</th>
-                                <th scope="col" className='text-uppercase'>Acciones</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {usuarios.map((usuario, index) => (
-                                <tr key={index}>
-                                    <td scope="row" className='filas'>{usuario.nombre}</td>
-                                    <td scope="row" className='filas'>{usuario.email}</td>
-                                    <td scope="row" className='filas'>{usuario.cargo}</td>
-                                    <td scope="row" className='filas'>{usuario.rol}</td>
-                                    <td scope="row" className='filas'>
-                                        <button onClick={() => handleEditarUsuario(usuario.uid)} className="editar-button">Editar</button>
-                                        <button onClick={() => handleEliminarUsuario(usuario.uid)} className="eliminar-button">Eliminar</button>
-                                    </td>
+                    <div className="table-responsive">
+                        <table className="table table-sm table-hover">
+                            <thead>
+                                <tr>
+                                    <th scope="col" className='text-uppercase'>Nombre</th>
+                                    <th scope="col" className='text-uppercase'>Correo</th>
+                                    <th scope="col" className='text-uppercase'>Cargo</th>
+                                    <th scope="col" className='text-uppercase'>Rol</th>
+                                    <th scope="col" className='text-uppercase'>Acciones</th>
                                 </tr>
-                            ))}
-                        </tbody>
-                    </table>
+                            </thead>
+                            <tbody>
+                                {usuarios.map((usuario, index) => (
+                                    <tr key={index}>
+                                        <td scope="row" className='filas'>{usuario.nombre}</td>
+                                        <td scope="row" className='filas'>{usuario.email}</td>
+                                        <td scope="row" className='filas'>{usuario.cargo}</td>
+                                        <td scope="row" className='filas'>{usuario.rol}</td>
+                                        <td scope="row" className='filas'>
+                                            <div >
+                                                <div class="row g-0 ">
+                                                    <div class="col-md-6">
+                                                        <button onClick={() => handleEditarUsuario(usuario.uid)} className="btn btn-editar">Editar</button>
+                                                    </div>
+                                                    <div class="col-md-6">
+                                                        <button onClick={() => handleEliminarUsuario(usuario.uid)} className="btn btn-eliminar">Eliminar</button>
+                                                    </div>
+                                                </div>
+                                            </div>
+
+
+                                        </td>
+                                    </tr>
+                                ))}
+                            </tbody>
+                        </table>
+                    </div>
+
                 </div>
             </div>
             <Footer />
